Refactor Omni router core structure
