+++ sglang_omni_router/__init__.py
@@ -1,4 +1,5 @@
-"""External HTTP router for complete SGLang-Omni server replicas."""
+# SPDX-License-Identifier: Apache-2.0
+"""External HTTP router for complete SGLang-Omni V1 server replicas."""
 
 from sglang_omni_router.config import RouterConfig, WorkerConfig
