--- sglang_omni_router/logging.py
@@ -1,29 +1,0 @@
-from __future__ import annotations
-
-import json
-import threading
-from datetime import datetime, timezone
-from pathlib import Path
-from typing import Any
-
-
-class RouteLogger:
-    def __init__(self, path: str | None) -> None:
-        self._path = Path(path) if path else None
-        self._lock = threading.Lock()
-        if self._path is not None:
-            self._path.parent.mkdir(parents=True, exist_ok=True)
-
-    def write(self, event: dict[str, Any]) -> None:
-        if self._path is None:
-            return
-
-        record = {
-            "timestamp": datetime.now(timezone.utc).isoformat(),
-            **event,
-        }
-        line = json.dumps(record, sort_keys=True)
-        with self._lock:
-            with self._path.open("a", encoding="utf-8") as f:
-                f.write(line)
-                f.write("\n")
