--- sglang_omni_router/serve/__init__.py
@@ -1,1 +1,0 @@
-"""Router serving entrypoints."""
