--- sglang_omni_router/serve/omni_router.py
@@ -1,6 +1,0 @@
-from __future__ import annotations
-
-from sglang_omni_router.launcher import main
-
-if __name__ == "__main__":
-    main()
